+++ webapp/src/App.tsx
@@ -41,6 +41,7 @@ const AdminReleaseEditor = lazy(() => import("./pages/AdminReleaseEditor"));
 const Files = lazy(() => import("./pages/Files"));
 const Tools = lazy(() => import("./pages/Tools"));
 const Store = lazy(() => import("./pages/Store"));
+const Samples = lazy(() => import("./pages/Samples"));
 const Dig = lazy(() => import("./pages/Dig"));
 const Lab = lazy(() => import("./pages/Lab"));
 const LabApp = lazy(() => import("./pages/LabApp"));
@@ -261,6 +262,7 @@ function App() {
 
             {/* Store */}
             <Route path="/store" element={<ModuleGuard flag="hideStore"><Store /></ModuleGuard>} />
+            <Route path="/samples" element={<ModuleGuard flag="hideSamples"><Samples /></ModuleGuard>} />
             <Route path="/dig" element={<ModuleGuard flag="hideDig"><Dig /></ModuleGuard>} />